parfectly working and remin some function to upgrade it

+++ admin/functions.php
@@ -1,4 +1,34 @@
 <?php 
+ function users_online(){
+     if (isset($_GET['onlineusers'])) {
+
+     global $connection;
+
+     if(!$connection) {
+         session_start();
+         include ("../includes/db.php");
+ $session = session_id();
+ $time = time();
+ $time_out_in_seconds = 5;
+ $time_out = $time - $time_out_in_seconds;
+ $query = "SELECT * FROM users_online WHERE session = '$session' ";
+ $send_query = mysqli_query($connection, $query);
+ $count = mysqli_num_rows($send_query);
+
+        if($count == NULL) {
+            mysqli_query($connection, "INSERT INTO users_online(session, time) VALUE('$session', '$time')");
+        } else {
+            mysqli_query($connection, "UPDATE users_online SET time = '$time' WHERE session = '$session'");
+        }
+        $users_online_query = mysqli_query($connection, "SELECT * FROM users_online WHERE time > '$time_out'");
+        echo $count_user = mysqli_num_rows($users_online_query);
+        }
+    } // get request isset
+}
+users_online();
+?>
+
+<?php 
  function confirmQuery($result){
      global $connection;
     if (!$result){
@@ -29,8 +59,8 @@ function insert_categories() {
 }
 ?>
 
-<!--  FIND ALL CATEGORY QUERY -->
 <?php
+// FIND ALL CATEGORY QUERY
 function findAllCategories(){
   global $connection;
     $query = "SELECT * FROM categories";
@@ -41,7 +71,7 @@ function findAllCategories(){
         echo "<tr>";
         echo "<td>{$cat_id}</td>";
         echo "<td>{$cat_title}</td>";
-        echo "<td><a href='categories.php?delete={$cat_id}'>Delete</a></td>";
+        echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete this category'); \" href='categories.php?delete={$cat_id}'>Delete</a></td>";
         echo "<td><a href='categories.php?edit={$cat_id}'>Edit</a></td>";
         echo "</tr>";
     }
@@ -50,8 +80,8 @@ function findAllCategories(){
 
 ?>
 
- <!-- DELETE QUERY -->
 <?php 
+ //DELETE QUERY
 function deleteCategories(){
 global $connection;
 
@@ -64,3 +94,4 @@ global $connection;
 }
 
 ?>
+
