+++ admin/includes/view_all_users.php
@@ -57,7 +57,7 @@
                 echo "<td><a href='users.php?admin=$user_id'>Admin</a></td>";
                 echo "<td><a href='users.php?subscriber=$user_id'>Subscriber</a></td>";
                 echo "<td><a href='users.php?source=edit_user&edit_user={$user_id}'>Edit</a></td>";
-                echo "<td><a href='users.php?delete=$user_id'>Delete</a></td>";
+                echo "<td><a onClick=\"javascript: return confirm('Are you sure you want to delete this user'); \" href='users.php?delete=$user_id'>Delete</a></td>";
                 echo "</tr>";
             }
         ?>
@@ -78,7 +78,7 @@
     if(isset($_GET['subscriber'])){
         $the_user_id = $_GET['subscriber'];
 
-    $query = "UPDATE users SET user_role= 'subscriberd' WHERE user_id = $the_user_id ";
+    $query = "UPDATE users SET user_role= 'subscriber' WHERE user_id = $the_user_id ";
     $change_to_sub_query = mysqli_query($connection, $query); 
     header("Location: users.php");
     }
